To update plugin to use and support java-lib plugin which provide api config - To remove compile configuration from all tests 1/2 - To replace compile configuration with api configuration - To fix findScalaVersions logic for main configs - Refactor DependencyInsights1 - Add tests for api configuration - Remove some deprecated methods - Manually wire api configuration for crossbuild sourcesets -
