+++ src/test/groovy/com/github/prokod/gradle/crossbuild/CrossBuildPluginImplementationMultiModuleTest.groovy
@@ -29,6 +29,7 @@ class CrossBuildPluginImplementationMultiModuleTest extends CrossBuildGradleRunn
     File propsFile
     File buildFile
     File libBuildFile
+    File libJavaFile
     File libScalaFile
     File lib2BuildFile
     File lib2ScalaFile
@@ -43,6 +44,7 @@ class CrossBuildPluginImplementationMultiModuleTest extends CrossBuildGradleRunn
         propsFile = file('gradle.properties')
         buildFile = file('build.gradle')
         libBuildFile = file('lib/build.gradle')
+        libJavaFile = file('lib/src/main/java/HelloWorldLibImpl.java')
         libScalaFile = file('lib/src/main/scala/Factorial.scala')
         lib2BuildFile = file('lib2/build.gradle')
         lib2ScalaFile = file('lib2/src/main/scala/CompileTimeFactorial.scala')
@@ -51,6 +53,566 @@ class CrossBuildPluginImplementationMultiModuleTest extends CrossBuildGradleRunn
         lib3JavaFile = file('lib3/src/main/java/CompileTimeFactorialUtils.java')
         appBuildFile = file('app/build.gradle')
         appScalaFile = file('app/src/main/scala/Test.scala')
+    }
+
+    /**
+     * Test Properties:
+     * <ul>
+     *     <li>Plugin apply mode: Eager</li>
+     *     <li>Running tasks that triggers main sourceset configurations: No</li>
+     * </ul>
+     * Leveraging layout 2 of propagating cross build configs to sub projects
+     * Layout 2 means:
+     * <ul>
+     *     <li>root-project</li>
+     *     <ul>
+     *         <li>{@code plugins} DSL for crossbuild plugin</li>
+     *         <li>{@code subprojects} block with {@code apply plugin:} {@link CrossBuildPlugin} followed by crossbuld plugin DSL</li>
+     *     </ul>
+     *     <li>sub-projects</li>
+     *     <ul>
+     *         <li>Nothing special</li>
+     *     </ul>
+     * </ul>
+     * NOTES:
+     * <ul>
+     *     <li>Gradle build is called in this test with specific cross compile tasks only: 'crossBuildSpark160_210Jar', 'crossBuildSpark240_211Jar'.
+     *     As a result, the main sourceset classpath configuration is not triggered to get resolved. Usual issues like scala-library is missing in 'app' module
+     *     than do not pop up and so the 'user' do not need to use api configuration for implicit scala-library dependency inclusion, or to explicitly include
+     *     scala-library dependency in all modules using implementation configuration as the plugin takes care of it for cross compile configurations.</li>
+     * </ul>
+     * @return
+     */
+    @Unroll
+    def "[gradle:#gradleVersion | default-scala-version:#defaultScalaVersion] applying crossbuild plugin on a multi-module project and calling crossBuildXXXJar tasks on it should build correctly"() {
+        given:
+        // root project settings.gradle
+        settingsFile << """
+include 'lib'
+include 'app'
+"""
+
+        buildFile << """
+plugins {
+    id 'com.github.prokod.gradle-crossbuild' apply false
+}
+
+allprojects {
+    group = 'com.github.prokod.it'
+    version = '1.0-SNAPSHOT'
+    
+    repositories {
+        mavenCentral()
+    }
+}
+
+subprojects {
+    apply plugin: 'com.github.prokod.gradle-crossbuild'
+    apply plugin: 'maven-publish'
+
+    crossBuild {
+        
+        scalaVersionsCatalog = ['2.11':'2.11.11']
+
+        builds {
+            spark160_210 
+            spark240_211
+        }
+    }
+    
+    publishing {
+        publications {
+            crossBuildSpark160_210(MavenPublication) {
+                ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
+                    artifact crossBuildSpark160_210Jar
+                ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
+            }
+            crossBuildSpark240_211(MavenPublication) {
+                ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
+                    artifact crossBuildSpark240_211Jar
+                ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
+            }
+        }
+    }
+
+    tasks.withType(GenerateMavenPom) { t ->
+        if (t.name.contains('CrossBuildSpark160_210')) {
+            t.destination = file("\$buildDir/generated-pom_2.10.xml")
+        }
+        if (t.name.contains('CrossBuildSpark240_211')) {
+            t.destination = file("\$buildDir/generated-pom_2.11.xml")
+        }
+    }
+}
+"""
+
+        libScalaFile << """
+import org.scalatest._
+
+trait HelloWorldLibApi {
+   def greet()
+}
+"""
+
+        libJavaFile << """
+
+public class HelloWorldLibImpl implements HelloWorldLibApi {
+   public void greet() {
+      System.out.println("Hello, world!");
+   }
+}
+"""
+
+        libBuildFile << """
+sourceSets {
+    main {
+        scala {
+            srcDirs = ['src/main/scala', 'src/main/java']
+        }
+        java {
+            srcDirs = []
+        }
+    }
+}
+
+dependencies {
+    implementation "org.scalatest:scalatest_?:3.0.1"
+    implementation "com.google.guava:guava:18.0"
+    implementation "org.scala-lang:scala-library:${defaultScalaVersion}.+"
+}
+"""
+
+        appScalaFile << """
+import HelloWorldLibImpl._
+
+object HelloWorldApp {
+   def main(args: Array[String]) {
+      new HelloWorldLibImpl().greet()
+   }
+}
+"""
+
+        appBuildFile << """
+dependencies {
+    implementation project(':lib')
+}
+"""
+
+        when:
+        Assume.assumeTrue(testMavenCentralAccess())
+        def result = GradleRunner.create()
+                .withGradleVersion(gradleVersion)
+                .withProjectDir(dir.root)
+                .withPluginClasspath()
+                .withDebug(true)
+                .withArguments('crossBuildSpark160_210Jar', 'crossBuildSpark240_211Jar', '--info', '--stacktrace')
+                .build()
+
+        then:
+        result.task(":lib:crossBuildSpark160_210Jar").outcome == SUCCESS
+        result.task(":lib:crossBuildSpark240_211Jar").outcome == SUCCESS
+        result.task(":app:crossBuildSpark160_210Jar").outcome == SUCCESS
+        result.task(":app:crossBuildSpark240_211Jar").outcome == SUCCESS
+
+        fileExists("$dir.root.absolutePath/lib/build/libs/lib_2.10*.jar")
+        fileExists("$dir.root.absolutePath/lib/build/libs/lib_2.11*.jar")
+        fileExists("$dir.root.absolutePath/app/build/libs/app_2.10*.jar")
+        fileExists("$dir.root.absolutePath/app/build/libs/app_2.11*.jar")
+
+        where:
+        gradleVersion   | defaultScalaVersion
+        '4.10.3'        | '2.12'
+        '5.6.4'         | '2.12'
+        '6.9.2'         | '2.10'
+        '7.3.3'         | '2.11'
+    }
+
+    /**
+     * Leveraging layout 2 of propagating cross build configs to sub projects
+     * Layout 2 means:
+     * <ul>
+     *     <li>root-project</li>
+     *     <ul>
+     *         <li>{@code plugins} DSL for crossbuild plugin</li>
+     *
+     *  <li>Running tasks that triggers main sourceset configurations: No</li>
+     *         <li>{@code subprojects} block with {@code apply plugin:} {@link CrossBuildPlugin} followed by crossbuld plugin DSL</li>
+     *     </ul>
+     *     <li>sub-projects</li>
+     *     <ul>
+     *         <li>Nothing special</li>
+     *     </ul>
+     * </ul>
+     *
+     * Expects to throw exception regarding org.gradle.api.CircularReferenceException: Circular dependency between lib and app
+     *
+     * @return
+     */
+    @Unroll
+    def "[gradle:#gradleVersion | default-scala-version:#defaultScalaVersion] applying crossbuild plugin on a multi-module project with circular dependency should not stack overflow  <git issue #72>)"() {
+        given:
+        // root project settings.gradle
+        settingsFile << """
+include 'lib'
+include 'app'
+"""
+
+        buildFile << """
+plugins {
+    id 'com.github.prokod.gradle-crossbuild' apply false
+}
+
+allprojects {
+    group = 'com.github.prokod.it'
+    version = '1.0-SNAPSHOT'
+    
+    repositories {
+        mavenCentral()
+    }
+}
+
+subprojects {
+    apply plugin: 'com.github.prokod.gradle-crossbuild'
+
+    crossBuild {
+        
+        scalaVersionsCatalog = ['2.11':'2.11.11']
+
+        builds {
+            spark160_210 
+            spark240_211
+        }
+    }
+}
+"""
+
+        libScalaFile << """
+import org.scalatest._
+
+trait HelloWorldLibApi {
+   def greet()
+}
+"""
+
+        libJavaFile << """
+
+public class HelloWorldLibImpl implements HelloWorldLibApi {
+   public void greet() {
+      System.out.println("Hello, world!");
+   }
+}
+"""
+
+        libBuildFile << """
+sourceSets {
+    main {
+        scala {
+            srcDirs = ['src/main/scala', 'src/main/java']
+        }
+        java {
+            srcDirs = []
+        }
+    }
+}
+
+dependencies {
+    implementation "org.scalatest:scalatest_?:3.0.1"
+    implementation "com.google.guava:guava:18.0"
+    implementation "org.scala-lang:scala-library:${defaultScalaVersion}.+"
+
+    implementation project(':app')
+}
+"""
+
+        appScalaFile << """
+import HelloWorldLibImpl._
+
+object HelloWorldApp {
+   def main(args: Array[String]) {
+      new HelloWorldLibImpl().greet()
+   }
+}
+"""
+
+        appBuildFile << """
+dependencies {
+    implementation project(':lib')
+}
+"""
+
+        when:
+        Assume.assumeTrue(testMavenCentralAccess())
+        def result = GradleRunner.create()
+                .withGradleVersion(gradleVersion)
+                .withProjectDir(dir.root)
+                .withPluginClasspath()
+                .withDebug(true)
+                .withArguments('crossBuildSpark160_210Jar', 'crossBuildSpark240_211Jar', '--info', '--stacktrace')
+                .build()
+
+        then:
+        org.gradle.testkit.runner.UnexpectedBuildFailure ex = thrown()
+        ex.buildResult.output.contains('org.gradle.api.CircularReferenceException: Circular dependency')
+
+        where:
+        gradleVersion   | defaultScalaVersion
+        '4.10.3'        | '2.12'
+        '5.6.4'         | '2.11'
+        '6.9.2'         | '2.11'
+        '7.3.3'         | '2.12'
+    }
+
+    @Unroll
+    def "[gradle:#gradleVersion | default-scala-version:#defaultScalaVersion] applying crossbuild plugin lazily on a multi-module project with dependency graph of depth 3 and with cross building dsl that is different on each submodule and with publishing dsl should produce expected: jars, pom files; and pom files content should be correct"() {
+        given:
+        // root project settings.gradle
+        settingsFile << """
+include 'lib'
+include 'lib2'
+include 'app'
+"""
+
+        buildFile << """
+plugins {
+    id 'com.github.prokod.gradle-crossbuild' apply false
+}
+
+allprojects {
+    group = 'com.github.prokod.it'
+    version = '1.0-SNAPSHOT'
+    
+    repositories {
+        mavenCentral()
+    }
+}
+
+subprojects {
+    project.pluginManager.withPlugin('com.github.prokod.gradle-crossbuild') {
+        if (!project.name.endsWith('app')) {
+            crossBuild {
+                scalaVersionsCatalog = ['2.12':'2.12.8']
+                builds {
+                    spark230_211 
+                    spark240_212
+                }
+            }
+        }
+    }
+    
+    project.pluginManager.withPlugin('maven-publish') {
+        if (!project.name.endsWith('app')) {
+            publishing {
+                publications {
+                    crossBuildSpark230_211(MavenPublication) {
+                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
+                            artifact crossBuildSpark230_211Jar
+                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
+                    }
+                    crossBuildSpark240_212(MavenPublication) {
+                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
+                            artifact crossBuildSpark240_212Jar
+                        ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
+                    }
+                }
+            }
+        }
+    }
+
+    tasks.withType(GenerateMavenPom) { t ->
+        if (t.name.contains('CrossBuildSpark230_211')) {
+            t.destination = file("\$buildDir/generated-pom_2.11.xml")
+        }
+        if (t.name.contains('CrossBuildSpark240_212')) {
+            t.destination = file("\$buildDir/generated-pom_2.12.xml")
+        }
+    }
+}
+"""
+
+        libScalaFile << """
+object Factorial {
+  // The actual implementation is regular old-fashioned scala code:
+  def normalFactorial(n: Int): Int =
+    if (n == 0) 1
+    else n * normalFactorial(n - 1)
+}
+"""
+
+        libBuildFile << """
+apply plugin: 'com.github.prokod.gradle-crossbuild'
+apply plugin: 'maven-publish'
+
+sourceSets {
+    main {
+        scala {
+            srcDirs = ['src/main/scala', 'src/main/java']
+        }
+        java {
+            srcDirs = []
+        }
+    }
+}
+
+dependencies {
+    implementation "com.google.guava:guava:18.0"
+    implementation "org.scala-lang:scala-library:${defaultScalaVersion}.+"
+}
+"""
+
+        lib2ScalaFile.withWriter('utf-8') {it.write """
+object CompileTimeFactorial {
+
+  import scala.language.experimental.macros
+  import Factorial._
+
+  // This function exposed to consumers has a normal Scala type:
+  def factorial(n: Int): Int =
+  // but it is implemented as a macro:
+  macro CompileTimeFactorial.factorial_impl
+
+  import scala.reflect.macros.blackbox.Context
+
+  // The macro implementation will receive a ‘Context’ and
+  // the AST’s of the parameters passed to it:
+  def factorial_impl(c: Context)(n: c.Expr[Int]): c.Expr[Int] = {
+    import c.universe._
+
+    // We can pattern-match on the AST:
+    n match {
+      case Expr(Literal(Constant(nValue: Int))) =>
+        // We perform the calculation:
+        val result = normalFactorial(nValue)
+        // And produce an AST for the result of the computation:
+        c.Expr(Literal(Constant(result)))
+      case other =>
+        // Yes, this will be printed at compile time:
+        println("Yow!")
+        ???
+    }
+  }
+}
+""" }
+
+        lib2BuildFile << """
+apply plugin: 'com.github.prokod.gradle-crossbuild'
+apply plugin: 'maven-publish'
+
+sourceSets {
+    main {
+        scala {
+            srcDirs = ['src/main/scala', 'src/main/java']
+        }
+        java {
+            srcDirs = []
+        }
+    }
+}
+
+dependencies {
+    implementation project(':lib')
+    implementation 'org.scala-lang:scala-reflect:2.12.8'
+    crossBuildSpark230_211Implementation 'org.scala-lang:scala-reflect:2.11.12'
+}
+"""
+
+        appScalaFile << """
+import CompileTimeFactorial._
+
+object Test extends App {
+    println(factorial(10))
+
+    // When uncommented, this will produce an error at compile-time, as we
+    // only implemented a case for an Int literal, not a variable:
+    // val n = 10
+    // println(factorial(n))
+}
+"""
+
+        appBuildFile << """
+apply plugin: 'com.github.prokod.gradle-crossbuild'
+apply plugin: 'maven-publish'
+
+crossBuild {
+    builds {
+        spark230_211 
+    }
+}
+
+publishing {
+    publications {
+        crossBuildSpark230_211(MavenPublication) {
+            ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : 'afterEvaluate {'}
+                artifact crossBuildSpark230_211Jar
+            ${publishTaskSupportingDeferredConfiguration(gradleVersion) ? '' : '}'}
+        }
+    }
+}
+        
+dependencies {
+    implementation project(':lib2')
+}
+"""
+
+        when:
+        Assume.assumeTrue(testMavenCentralAccess())
+        def result = GradleRunner.create()
+                .withGradleVersion(gradleVersion)
+                .withProjectDir(dir.root)
+                .withPluginClasspath()
+                .withDebug(true)
+                .withArguments('publishToMavenLocal', '--info', '--stacktrace')
+                .build()
+
+        then:
+        result.task(":app:crossBuildSpark230_211Jar").outcome == SUCCESS
+
+        fileExists("$dir.root.absolutePath/lib/build/libs/lib_2.11*.jar")
+        fileExists("$dir.root.absolutePath/lib/build/libs/lib_2.12*.jar")
+        fileExists("$dir.root.absolutePath/lib2/build/libs/lib2_2.11*.jar")
+        fileExists("$dir.root.absolutePath/lib2/build/libs/lib2_2.12*.jar")
+        fileExists("$dir.root.absolutePath/app/build/libs/app_2.11*.jar")
+        !fileExists("$dir.root.absolutePath/app/build/libs/app_2.12*.jar")
+
+        def pom211 = new File("${dir.root.absolutePath}${File.separator}lib${File.separator}build${File.separator}generated-pom_2.11.xml").text
+        def pom212 = new File("${dir.root.absolutePath}${File.separator}lib${File.separator}build${File.separator}generated-pom_2.12.xml").text
+
+        !pom211.contains('2.12.')
+        pom211.contains('2.11.12')
+        pom211.contains('18.0')
+        !pom212.contains('2.12.+')
+        !pom212.contains('2.11.')
+        pom212.contains('2.12.8')
+        pom212.contains('18.0')
+
+        def lib2pom211 = new File("${dir.root.absolutePath}${File.separator}lib2${File.separator}build${File.separator}generated-pom_2.11.xml").text
+        def lib2pom212 = new File("${dir.root.absolutePath}${File.separator}lib2${File.separator}build${File.separator}generated-pom_2.12.xml").text
+
+        !lib2pom211.contains('2.12.')
+        lib2pom211.contains('scala-reflect')
+        lib2pom211.contains('2.11.12')
+        lib2pom211.contains('1.0-SNAPSHOT')
+        lib2pom211.contains('lib_2.11')
+        !lib2pom212.contains('2.12.+')
+        !lib2pom212.contains('2.11.')
+        lib2pom212.contains('2.12.8')
+        lib2pom212.contains('scala-reflect')
+        lib2pom212.contains('1.0-SNAPSHOT')
+        lib2pom212.contains('lib_2.12')
+
+        def appPom211 = new File("${dir.root.absolutePath}${File.separator}app${File.separator}build${File.separator}generated-pom_2.11.xml").text
+        def appPom212Exist = new File("${dir.root.absolutePath}${File.separator}app${File.separator}build${File.separator}generated-pom_2.12.xml").exists()
+
+        !appPom211.contains('2.12.')
+        appPom211.contains('2.11.12')
+        appPom211.contains('lib2_2.11')
+        !appPom212Exist
+
+        where:
+        gradleVersion   | defaultScalaVersion
+        '4.10.3'        | '2.11'
+        '5.6.4'         | '2.12'
+        '6.9.2'         | '2.12'
+        '7.3.3'         | '2.11'
     }
 
     /**
@@ -157,7 +719,7 @@ dependencies {
 }
 """
 
-        lib2ScalaFile << """
+        lib2ScalaFile.withWriter('utf-8') {it.write """
 object CompileTimeFactorial {
 
   import scala.language.experimental.macros
@@ -189,7 +751,7 @@ object CompileTimeFactorial {
     }
   }
 }
-"""
+""" }
 
         lib2BuildFile << """
 sourceSets {
@@ -271,7 +833,7 @@ dependencies {
         def expectedJsonAsText = loadResourceAsText(dsv: defaultScalaVersion,
                 defaultOrRuntime: gradleVersion.startsWith('4') ? 'default' : 'runtime',
                 defaultOrCompile: gradleVersion.startsWith('4') ? 'default' : 'compile',
-                _2_11_12_: gradleVersion.startsWith('6') ? '-2.11.12' : '',
+                _2_11_12_: gradleVersion.startsWith('6') || gradleVersion.startsWith('7') ? '-2.11.12' : '',
                 '/04-app_builds_resolved_configurations.json')
         def appResolvedConfigurationReportFile = findFile("*/app_builds_resolved_configurations.json")
 
@@ -357,14 +919,15 @@ dependencies {
         gradleVersion | defaultScalaVersion
         '4.10.3'      | '2.12'
         '5.6.4'       | '2.11'
-        '6.5'         | '2.12'
-        '7.2'         | '2.11'
+        '6.9.2'       | '2.12'
+        '7.3.3'       | '2.11'
     }
 
     /**
      * Test Properties:
      * <ul>
      *     <li>Plugin apply mode: Lazy</li>
+     *     <li>Gradle compatibility matrix: 5.x, 6.x, 7.x</li>
      * </ul>
      * Here lib3 is a non cross build dependency.
      * This test checks that when lib3 is added to dependent app module, while using 'implementation' configuration,
@@ -440,6 +1003,7 @@ subprojects {
                     }
                 }
             }
+
         }
     }
 
@@ -602,9 +1166,8 @@ dependencies {
 
         where:
         gradleVersion   | defaultScalaVersion
-        '4.10.3'        | '2.11'
         '5.6.4'         | '2.12'
-        '6.5'           | '2.12'
-        '7.2'           | '2.11'
+        '6.9.2'         | '2.12'
+        '7.3.3'         | '2.11'
     }
 }
