+++ src/test/groovy/com/github/prokod/gradle/crossbuild/CrossBuildPluginTest.groovy
@@ -73,7 +73,7 @@ crossBuildResolvedDsl - Summary report for cross building resolved Dsl
         result.output.contains('_2.12]')
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -114,7 +114,7 @@ crossBuild {
         result.task(":tasks").outcome == SUCCESS
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -168,7 +168,7 @@ publishing {
         result.task(":tasks").outcome == SUCCESS
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -206,7 +206,8 @@ crossBuild {
         gradleVersion   | pluginId
         '4.10.3'        | 'com.github.prokod.gradle-crossbuild'
         '5.6.4'         | 'com.github.prokod.gradle-crossbuild-scala'
-        '6.5'           | 'com.github.prokod.gradle-crossbuild-scala'
+        '6.9.2'         | 'com.github.prokod.gradle-crossbuild-scala'
+        '7.3.3'         | 'com.github.prokod.gradle-crossbuild-scala'
     }
 
     @Unroll
@@ -247,7 +248,7 @@ crossBuild {
         result.task(":crossBuildV211Jar").outcome == SUCCESS
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -296,7 +297,7 @@ gradle.projectsEvaluated {
         result.task(":crossBuildV211Jar").outcome == SUCCESS
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -384,7 +385,7 @@ gradle.projectsEvaluated {
         result.task(":crossBuildSpark24_212Jar").outcome == SUCCESS
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -454,7 +455,7 @@ crossBuildV213Jar - Assembles a jar archive containing 213 classes
         result.task(":tasks").outcome == SUCCESS
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -513,7 +514,7 @@ publishing {
         result.task(":publishCrossBuildV211PublicationToMavenLocal").outcome == SUCCESS
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -561,7 +562,7 @@ sourceSets {
         result.task(":crossBuild211Jar").outcome == SUCCESS
 
         where:
-        gradleVersion << ['2.14.1', '3.0', '4.1']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -594,7 +595,7 @@ crossBuild {
         thrown(RuntimeException)
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -629,7 +630,7 @@ crossBuild {
         thrown(RuntimeException)
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 
     @Unroll
@@ -664,6 +665,6 @@ crossBuild {
         result.task(":crossBuildV213Jar").outcome == SUCCESS
 
         where:
-        gradleVersion << ['4.10.3', '5.6.4', '6.5']
+        gradleVersion << ['4.10.3', '5.6.4', '6.9.2', '7.3.3']
     }
 }
